Fix the TreeView bug where a deleted city was not removed from the window.

+++ Assets/SEE/Controls/Actions/ContextMenuAction.cs
@@ -421,8 +421,11 @@ namespace SEE.Controls.Actions
                 entries.Add(new PopupMenuHeading("Target: " + target, Priority: int.MaxValue));
             }
 
-            entries.Add(new PopupMenuAction(graphElement is Node no && no.IsArchitectureOrImplementationRoot() ?
-                "Clear" : "Delete", () => DeleteElement().Forget(), Icons.Trash, Priority: 0));
+            if (appendActions == null || !appendActions.Any(action => action.Name.Contains("Delete")))
+            {
+                entries.Add(new PopupMenuAction(graphElement is Node no && no.IsArchitectureOrImplementationRoot() ?
+                    "Clear" : "Delete", () => DeleteElement().Forget(), Icons.Trash, Priority: 0));
+            }
 
             entries.Add(new PopupMenuActionDoubleIcon("Inspect", () =>
             {
@@ -465,12 +468,7 @@ namespace SEE.Controls.Actions
             {
                 if (graphElement is Node node && node.IsRoot())
                 {
-                    string deleteMessage = $"Do you really want to delete the city?\r\nThis action cannot be undone.";
-                    if (await ConfirmDialog.ConfirmAsync(ConfirmConfiguration.Delete(deleteMessage)))
-                    {
-                        GameElementDeleter.DeleteRoot(node.GameObject());
-                        new DeleteNetAction(graphElement.ID).Execute();
-                    }
+                    DeleteCityAsync(node).Forget();
                     return;
                 }
                 if (gameObject != null)
@@ -935,6 +933,28 @@ namespace SEE.Controls.Actions
         {
             LocalPlayer.TryGetPlayerMenu(out PlayerMenu menu);
             menu.UpdateActiveEntry();
+        }
+
+        /// <summary>
+        /// Deletes the root of a city.
+        /// This cannot be undone, so a <see cref="ConfirmDialog"/> is opened first to obtain
+        /// confirmation for the deletion.
+        /// </summary>
+        /// <param name="node">The root node that is to be deleted.</param>
+        /// <param name="action">An action to be executed after the deletion. Can be null</param>
+        /// <returns>Waits until a response from the <see cref="ConfirmDialog"/> is received.</returns>
+        public async static UniTask DeleteCityAsync(Node node, Action action = null)
+        {
+            if (node.IsRoot())
+            {
+                string deleteMessage = $"Do you really want to delete the city?\r\nThis action cannot be undone.";
+                if (await ConfirmDialog.ConfirmAsync(ConfirmConfiguration.Delete(deleteMessage)))
+                {
+                    GameElementDeleter.DeleteRoot(node.GameObject());
+                    new DeleteNetAction(node.ID).Execute();
+                    action?.Invoke();
+                }
+            }
         }
     }
 }
